[7185] Amended trainee filter params attributes to be versioned
