+++ spec/models/api/attributes_spec.rb
@@ -3,7 +3,7 @@
 require "rails_helper"
 
 describe Api::Attributes do
-  attributes = %i[degree hesa_trainee_detail nationality placement trainee map_hesa withdrawal]
+  attributes = %i[degree hesa_trainee_detail nationality placement trainee map_hesa withdrawal trainee_filter_params]
   context "v0.1" do
     attributes.each do |attribute|
       it "#{attribute} has been implemented" do
